Move DgmlWriter to Dgml subfolder
