+++ src/NRules.Debugger.Visualizer/NRules.Debugger.Visualizer.DebuggeeSide/SessionPerformanceObjectSource.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 using Microsoft.VisualStudio.DebuggerVisualizers;
-using NRules.Diagnostics;
+using NRules.Diagnostics.Dgml;
 
 namespace NRules.Debugger.Visualizer
 {
